Change geometry and layout paradigm for HLMenuNode.
This change is BACKWARDS INCOMPATIBLE with the previous version.

Previously, the menu node did not track its overall size, and so did
not offer size or anchorPoint properties.  Furthermore, it always laid
out the first button centered on (0, 0).  And finally, it offered a
single button-spacing option that didn't account for the size of the
button.

This had a number of undesirable properties.  First, it was hard to
maintain the correct visual position of the menu node when the number
of buttons in the current menu changed.  Second, it was hard to
position the menu node according the size of the buttons.  (Consider,
for example, trying to position the menu node at the top of the
screen: The first button displayed would be halfway off the screen to
the top.)

The new layout system offers more-useful geometric properties: The new
item spacing option is a separator size between the edges of the
buttons, and the menu node now offers size and anchorPoint properties.
(The buttons are still always horizontally centered within the menu
node, but a justification option could easily be added if desired.)