Remove hide/show functionality from HLToolbarNode.
This functionality belongs in the owner or controller, and not the
node itself.  In particular, the owner is responsible for adding the
toolbar node as a child, positioning it, and scaling it.  Arguably the
toolbar node could provide the SKAction animation sequence as a
convenience for multiple controllers reuse the same show/hide
behavior, but in practice each controller wants to do it slightly
differently, and so there is no point.  Furthermore, such show/hide
functionality could be shared not just between different controllers,
but in fact also shared by different scene components, and so again,
the code does not belong in a particular node.