Work around undocumented size method in SKNode.
As of iOS 10, there is an undocumented `size` method exposed in
SKNode.  For SKLabelNode it returns 0, which caused troubles for my
code trying to automatically determine the size of an unidentified
node (based on whether it responded to a selector or not).  The
workaround is to check for SKLabelNode in particular before checking
for the size selector.