Improve interface and documentation for HLAction.
The old use case was to provide encodable alternatives to SKActions.
The new perhaps more-common use case is to provide perform-selector
actions that don't retain their target strongly.  Added the
no-argument HLPerformSelectorWeak for this purpose, and at the same
time realized that a drop-in HLAction class method interface (parallel
to SKAction class method interface) would be pretty sweet.