Check return type of "size" methods before invoking under macOS.
In a few places in HLSpriteKit, generic SKNodes are tested to see if
they respond to a "size" or "setSize" selector.  Under macOS, the
compiler finds a number of size methods which return NSSize rather
than CGSize.  This might be some kind of fixable typedef or macro
confusion, but this diff implements a conservative fix: At runtime,
get the method signature of the size method for the SKNode object, and
if it returns a CGSize type, then use it; otherwise, don't use it.