Replace `HLLabelNodeVerticalAlignmentMode` with `HLLabelHeightMode`.
Also add a new scene to Example to demonstrate alignment options.  The
alignment code has been reparameterized so that in place of a single
combo mode there are now two separate modes: an alignment mode and a
height mode.  The same results should be obtainable by mapping the old
`HLLabelNodeVerticalAlignmentMode` to an equivalent
`HLLabelHeightMode` and passing `SKLabelVerticalAlignmentModeCenter`
as the new alignment mode.