Ease the default easing function curves in HLAction.
Default ease-out for SpriteKit SKAction (SKActionTimingEaseOut) uses a
cubic Bezier with control points (0,0) (1/3,1/3) (2/3,1) and (1,1).
This is much gentler than the HLAction default, which was using cubic
polynomial easing.  This diff replaces the HLAction default with sine
easing, which is still more aggressive than SKAction, but is closer,
and is easier to calculate.