Add DEBUG-mode detection for reuse of HLAction.
SKActions are immutable and can be (and should be) reused.  HLActions
are stateful, and so should only be used once.  When using the latter
as a replacement for the former, it's a natural mistake to make.
Detecting the problem is very possible, but requires a new state
variable in HLAction; since it's mostly a development/debug-only
problem, for now the new error detection and state are DEBUG-only.