Allow pans and swipes to fall through to lower targets by default.
This is a major change with the way gesture targets work by default.
Previously it was up to each gesture target to decide which types of
gestures would "fall through" to other targets (if unhandled).  But
the same logic kept repeating with all targets: Taps and long-presses
should be blocked by the top target (whether handled or not); all
other gestures should not be blocked by the top target (if unhandled).