Use SKAction rather than NSObject callback mechanism for scrolling.
The SKAction mechanism is better for a couple (theoretical?) reasons.
First, it presumably be more careful to call back on the main thread
at an appropriate time in the SKScene frame cycle.  Second, it allows
explicit creation of a weak reference, so that (in this case) the
HLScrollNode is not retained by the callback.