Move logic for absorbing gestures into each gesture target.
Previously, each gesture target decided whether a particular gesture
was "inside" itself or not, and then the controller (HLScene) decided
whether that meant the gesture was absorbed.  (In particular, taps and
long-presses would always be absorbed if "inside" a gesture target.)
Now, each gesture target decides what gestures it absorbs.